TelemetryPartA->TelemetryItem, Resource to BaseExporter
