+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Tests/AzureMonitorTestTraceExporter.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using OpenTelemetry;
+
+namespace Azure.Monitor.OpenTelemetry.Exporter
+{
+    internal class AzureMonitorTestTraceExporter<Activity> : AzureMonitorBaseExporter<Activity>
+        where Activity : class
+    {
+        public override ExportResult Export(in Batch<Activity> batch)
+        {
+            return ExportResult.Success;
+        }
+    }
+}
